Rename c01000 variable to loss_limited_net_capital_gains
