+++ policyengine_us/parameters/gov/irs/investment/income/sources.yaml
@@ -4,7 +4,7 @@ values:
     - taxable_interest_income
     - dividend_income
     - rental_income
-    - c01000  # Loss-limited capital gains.
+    - loss_limited_net_capital_gains
 metadata:
   unit: currency-USD
   label: net investment income sources
