+++ policyengine_us/tests/policy/baseline/gov/states/nj/tax/income/credits/eitc/nj_eitc_income_eligible.yaml
@@ -13,7 +13,7 @@
   absolute_error_margin: 0
   input:
     adjusted_gross_income: 22_600
-    capital_gains: 11_000
+    short_term_capital_gains: 11_000
     filing_status: JOINT
     state_code: NJ
   output:
