+++ policyengine_us/variables/gov/states/nj/tax/income/credits/eitc/nj_eitc_income_eligible.py
@@ -14,18 +14,15 @@ class nj_eitc_income_eligible(Variable):
         p = parameters(period).gov.irs.credits.eitc
 
         # Check if they are above the investment income limit.
-        no_loss_capital_gains = max_(
-            0,
-            add(tax_unit, period, ["capital_gains"]),
-        )
+        no_loss_capital_gains = max_(0, tax_unit("net_capital_gains", period))
         eitc_investment_income = (
             add(
                 tax_unit,
                 period,
                 ["net_investment_income", "tax_exempt_interest_income"],
             )
-            # Replace limited-loss capital gains with no-loss capital gains.
-            - tax_unit("c01000", period)  # Limited-loss capital gains.
+            # replace limited-loss capital gains with no-loss capital gains
+            - tax_unit("loss_limited_net_capital_gains", period)
             + no_loss_capital_gains
         )
         inv_income_disqualified = (
